Move eclipse formatter tests to a single integration test with before/after
