+++ gradle-baseline-java/build.gradle
@@ -120,3 +120,9 @@ tasks.publish.dependsOn tasks.publishPlugins
 publishPlugins.onlyIf { versionDetails().isCleanTag }
 project.ext.'gradle.publish.key' = System.env["GRADLE_KEY"]
 project.ext.'gradle.publish.secret' = System.env["GRADLE_SECRET"]
+
+// Run `./gradlew test -Drecreate=true` to recreate all the expected
+// generated code that we have checked into the repo.
+tasks.withType(Test) {
+    systemProperty 'recreate', System.getProperty('recreate', 'false')
+}
