+++ gradle-baseline-java/src/test/resources/com/palantir/baseline/formatter-expected/.baseline/copyright/999_palantir.txt
@@ -1,0 +1,1 @@
+(c) Copyright ${today.year} Palantir Technologies Inc. All rights reserved.
