+++ gradle-baseline-java/src/test/resources/com/palantir/baseline/formatter-expected/build.gradle
@@ -1,0 +1,1 @@
+buildscript { repositories { mavenCentral() }}
