+++ gradle-baseline-java/src/test/resources/com/palantir/baseline/formatter-expected/java/src/main/java/test/BlankLines.java
