+++ gradle-baseline-java/src/test/resources/com/palantir/baseline/formatter-expected/java/src/main/java/test/Invalid.java
@@ -1,0 +1,12 @@
+package test;
+
+public class Test {
+    void test() {
+        int x = 1;
+        System.out.println(
+                "Hello");
+        Optional.of("hello").orElseGet(() -> {
+            return "Hello World";
+        });
+    }
+}
