+++ gradle-baseline-java/src/test/resources/com/palantir/baseline/formatter-in/build.gradle
@@ -1,0 +1,1 @@
+buildscript { repositories { mavenCentral() }}
