+++ gradle-baseline-java/src/test/resources/com/palantir/baseline/formatter-in/java/src/foo/java/test/Invalid.java
@@ -1,0 +1,10 @@
+package test;
+import com.java.unused;
+public class Test { void test() {int x = 1;
+    System.out.println(
+            "Hello"
+    );
+    Optional.of("hello").orElseGet(() -> {
+        return "Hello World";
+    });
+} }
