+++ gradle-baseline-java/src/test/resources/com/palantir/baseline/formatter-in/java/src/main/java/test/BlankLines.java
