+++ gradle-baseline-java/src/test/resources/com/palantir/baseline/formatter-in/settings.gradle
@@ -1,0 +1,1 @@
+include 'java'
